Change from cat to echo to add user.
It's not need to add a complete templates of passwd, group and shadow since the debootstrap will add them just fine. Change from cat to echo to add user in the passwd, group and shadow. It's more simpler that way. Also we don't have to edit the script every time in future if Debian add more new users or group.

Move from before 2nd stage to after 2nd stage so it will add to exist files instead of create new files.